+++ src/Strategy/CalcularPoderEnergy.java
@@ -3,8 +3,8 @@ package Strategy;
 public class CalcularPoderEnergy implements EstrategiaCalcularPoder {
 
 	@Override
-	public double calcularPoder(String datos) {
-		return (double) Integer.parseInt(datos);
+	public int calcularPoder(String datos) {
+		return  Integer.parseInt(datos);
 	}
 
 }
