+++ src/Strategy/CalcularPoderItem.java
@@ -3,8 +3,8 @@ package Strategy;
 public class CalcularPoderItem implements EstrategiaCalcularPoder {
 
 	@Override
-	public double calcularPoder(String datos) {
-		return (double) Integer.parseInt(datos)*20;
+	public int calcularPoder(String datos) {
+		return  Integer.parseInt(datos)*20;
 	}
 	
 }
